# Intermediate Point DCEL Optimisation