--- a/geom/alg_binary_op.go
+++ b/geom/alg_binary_op.go
@@ -92,6 +92,7 @@ func createOverlay(a, b Geometry) (*doublyConnectedEdgeList, error) {
 
 	dcelA := newDCELFromGeometry(a, ghosts, inputAMask, interactionPoints)
 	dcelB := newDCELFromGeometry(b, ghosts, inputBMask, interactionPoints)
+
 	err = dcelA.overlay(dcelB)
 	return dcelA, err
 }