--- a/geom/dcel_edge_set.go
+++ b/geom/dcel_edge_set.go
@@ -7,7 +7,7 @@ package geom
 // the edge, the second is the first intermediate point (or a repeat of the
 // first XY if there are no intermediate points), and the third is the start
 // point of the next edge.
-type edgeSet map[[3]XY]struct{}
+type edgeSet map[[3]XY]*halfEdgeRecord
 
 func (s edgeSet) containsEdge(e *halfEdgeRecord) bool {
 	k := s.key(e.origin.coords, e.intermediate, e.next.origin.coords)
@@ -27,17 +27,35 @@ func (s edgeSet) containsStartIntermediateEnd(start XY, intermediate []XY, end X
 
 func (s edgeSet) insertEdge(e *halfEdgeRecord) {
 	k := s.key(e.origin.coords, e.intermediate, e.next.origin.coords)
-	s[k] = struct{}{}
+	s[k] = e
 }
 
 func (s edgeSet) insertLine(ln line) {
 	k := s.key(ln.a, nil, ln.b)
-	s[k] = struct{}{}
+	s[k] = nil // TODO: this is a bit weird...
 }
 
 func (s edgeSet) insertStartIntermediateEnd(start XY, intermediate []XY, end XY) {
 	k := s.key(start, intermediate, end)
-	s[k] = struct{}{}
+	s[k] = nil // TODO: this is a bit weird...
+}
+
+func (s edgeSet) lookupEdge(e *halfEdgeRecord) (*halfEdgeRecord, bool) {
+	k := s.key(e.origin.coords, e.intermediate, e.next.origin.coords)
+	e, ok := s[k]
+	return e, ok
+}
+
+func (s edgeSet) lookupLine(ln line) (*halfEdgeRecord, bool) {
+	k := s.key(ln.a, nil, ln.b)
+	e, ok := s[k]
+	return e, ok
+}
+
+func (s edgeSet) lookupStartIntermediateEnd(start XY, intermediate []XY, end XY) (*halfEdgeRecord, bool) {
+	k := s.key(start, intermediate, end)
+	e, ok := s[k]
+	return e, ok
 }
 
 func (edgeSet) key(start XY, intermediate []XY, end XY) [3]XY {