… geom/alg_binary_op_test.go
… geom/dcel_extract.go
… geom/dcel_interaction_points.go
… geom/dcel_interaction_points_test.go
… geom/dcel_overlay.go
… geom/dcel_re_noding_test.go
… geom/dcel_test.go
… geom/util.go
… internal/cmprefimpl/cmpgeos/checks.go