## Description

DECL Optimisation:

The idea:

- If you zoom in on a DCEL, many parts of it a just a fancy doubly linked list.
- There are only non-linear parts of the DCEL as very specific points (e.g. where two geometries interact, or a geometry interacts with itself).
- Because Go is a GC language, doubly linked lists are not great, especially if they contain 10s of thousands of nodes.
- The main insight is that we could just store parts of the DCEL that don't do anything fancy (the linear parts) as `[]XY` values on `halfEdgeRecord`s.
Changes:

- Add algorithm to determine "interaction" points. These are the points we will need to create new nodes (rather than continuing a `[]XY` sequence).
- Update DCEL constructors to handle intermediate nodes.
- Update overlay code to handle intermediate points. The main changes here are a few places where an edge set is created. There are also some changes around radial sorting of edges around vertices.
- Update DCEL -> Geometry conversion to include intermediate points in the final result.

## Benchmark Results:

Shows some decent improvement, most likely due to less allocations.

```
COMPARISON
name                  old time/op    new time/op    delta
Intersection/n=10        183µs ± 4%     141µs ± 5%  -22.91%  (p=0.000 n=14+15)
Intersection/n=100      1.46ms ± 2%    0.96ms ± 9%  -34.46%  (p=0.000 n=13+14)
Intersection/n=1000     17.6ms ± 2%     9.4ms ± 4%  -46.60%  (p=0.000 n=13+15)
Intersection/n=10000     234ms ± 6%     122ms ± 2%  -47.82%  (p=0.000 n=15+14)

name                  old alloc/op   new alloc/op   delta
Intersection/n=10       58.9kB ± 0%    47.4kB ± 0%  -19.39%  (p=0.000 n=15+15)
Intersection/n=100       473kB ± 0%     319kB ± 0%  -32.52%  (p=0.000 n=15+15)
Intersection/n=1000     5.86MB ± 0%    3.60MB ± 0%  -38.58%  (p=0.000 n=15+14)
Intersection/n=10000    56.1MB ± 0%    36.4MB ± 0%  -35.23%  (p=0.000 n=15+15)

name                  old allocs/op  new allocs/op  delta
Intersection/n=10          749 ± 0%       624 ± 0%  -16.69%  (p=0.000 n=15+15)
Intersection/n=100       5.18k ± 0%     3.74k ± 0%  -27.69%  (p=0.000 n=15+15)
Intersection/n=1000      49.4k ± 0%     35.0k ± 0%  -29.06%  (p=0.000 n=15+14)
Intersection/n=10000      492k ± 0%      348k ± 0%  -29.23%  (p=0.000 n=15+15)
```